actual-solution/requirements.txt
# Core Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# RAG & AI - Gemini Integration
chromadb>=0.4.15
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
langchain>=0.0.300
langchain-google-genai>=0.0.5
groq>=0.4.0

# Document Processing
pypdf2>=3.0.0
python-docx>=1.0.0
python-multipart>=0.0.6

# Database & Caching
sqlalchemy>=2.0.0
alembic>=1.12.0
redis>=5.0.0

# Utilities
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
loguru>=0.7.0

# Frontend
flask>=2.3.0
flask-socketio>=5.3.0
flask-cors>=4.0.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0